{{cookiecutter.project_name}}/requirements-dev.txt
-i https://pypi.org/simple
aiohttp==3.6.2
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==20.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
coverage==5.3
ddtrace==0.41.2
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==1.7.0; python_version < '3.8'
iniconfig==1.0.1
intervaltree==3.1.0
more-itertools==8.5.0; python_version >= '3.5'
multidict==4.7.6; python_version >= '3.5'
packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
protobuf==3.13.0
psycopg2==2.8.6
py==1.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.6.1; python_version >= '3.6'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest==6.0.2
python-dateutil==2.8.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sortedcontainers==2.2.2
sqlalchemy==1.3.19
starlette==0.13.6; python_version >= '3.6'
tenacity==6.2.0
toml==0.10.1
typing-extensions==3.7.4.3; python_version < '3.8'
uvicorn==0.11.8
uvloop==0.14.0; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
websockets==8.1; python_full_version >= '3.6.1'
yarl==1.5.1; python_version >= '3.5'
zipp==3.1.0; python_version >= '3.6'
